Fixed type issues with data_pool

+++ utility/data_pool.c
@@ -8,14 +8,14 @@
 /** init new data pool
  *  pre-mallocs a block of contiguous data
  */
-data_pool* init_data_pool(int element_size, int initial_size){
-    data_pool* new_pool = safe_malloc(sizeof(data_pool));
-    new_pool->free_list = create_stack(sizeof(void*));
+data_pool* init_data_pool(uint32_t element_size, uint32_t initial_size){
+    data_pool* new_pool = safe_malloc( (size_t) sizeof(data_pool));
+    new_pool->free_list = create_stack( (size_t) sizeof(void*));
     new_pool->element_size = element_size;
     new_pool->max = initial_size;
     new_pool->remaining = initial_size;
-    new_pool->data = safe_malloc(element_size * initial_size);
-    for(int i=0; i<initial_size; i++){
+    new_pool->data = safe_malloc( (size_t) element_size * initial_size);
+    for(uint32_t i=0; i<initial_size; i++){
         push(new_pool->free_list, new_pool->data + i * element_size);
     }
     return new_pool;
@@ -40,7 +40,7 @@ void reset_pool(data_pool* pool){
 
 void expand_data_pool(data_pool* pool){
     pool->max *= 2;
-    pool->data = safe_realloc(pool->data, pool->max * pool->element_size);
+    pool->data = safe_realloc(pool->data,  (size_t) pool->max * pool->element_size);
     pool->remaining = pool->max/2;
     for(uint32_t i=0; i<pool->max; i++){
         push(pool->free_list, pool->data + i * pool->element_size);
